native: DTA flow convergence + LP64 Symbol fixes
Remove gNativeVenueDir hack and NativeVenueInit() — venue rendering
now flows through proper DTA panel transitions. Remove manual venue
DrawShowing from App::RunWithoutDebugging.

Fix LP64 Symbol truncation: deprecate operator int() on native,
fix Locale::FindDataIndex to compare Str() pointers directly,
remove spurious (int) casts in ChooseModeProvider/PlaylistSongProvider.

Add UIScreen focus panel diagnostic, DTA flow test, session docs.

+++ src/lazer/meta_ham/ChooseModeProvider.cpp
@@ -144,7 +144,7 @@ RndMat *ChooseModeProvider::Mat(int, int i_iData, UIListMesh *mesh) const {
     } else {
         bool isOnePlusIcon = mesh->Matches("icon_1p_plus");
         if (isOnePlusIcon) {
-            if ((int)dataSym == custom_party) {
+            if (dataSym == custom_party) {
                 result = nullptr;
             } else {
                 result = mesh->DefaultMat();
